requirements.txt
spoon-ai-sdk>=0.3.0
spoon-toolkits>=0.2.0
pypdf>=3.0.0
requests>=2.31.0
anthropic>=0.34.0
openai>=1.0.0
python-dotenv>=1.0.0
networkx>=3.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
scipy>=1.10.0

